[mod] odysee: move format_duration helper into utils.py
